+++ src/ui/svelte/apps-menu/components/AllAppsView.svelte
@@ -1,20 +1,13 @@
 <script lang="ts">
-  import { type StartMenuItem } from "@seelen-ui/lib/types";
   import { t } from "../i18n";
   import AppItem from "./AppItem.svelte";
   import { searchState, getItemKey } from "../state/search.svelte";
-
-  interface Props {
-    onContextMenu: (event: MouseEvent, item: StartMenuItem) => void;
-  }
-
-  let { onContextMenu }: Props = $props();
 </script>
 
 <div class="all-apps-view">
   <div class="all-apps-view-list">
     {#each searchState.searchedItems as item, idx (getItemKey(item))}
-      <AppItem {item} {idx} {onContextMenu} lazy />
+      <AppItem {item} {idx} lazy />
     {/each}
   </div>