+++ src/ui/svelte/notifications/state.svelte.ts
@@ -1,10 +1,8 @@
-import { invoke, SeelenCommand, SeelenEvent, Settings, subscribe, Widget } from "@seelen-ui/lib";
+import { invoke, SeelenCommand, SeelenEvent, Settings, subscribe } from "@seelen-ui/lib";
 import type { AppNotification, NotificationsMode } from "@seelen-ui/lib/types";
 import { locale } from "./i18n/index.ts";
 import { writable } from "svelte/store";
 import { lazyRune } from "libs/ui/svelte/utils/LazyRune.svelte.ts";
-
-let widget = Widget.getCurrent();
 
 let settings = writable(await Settings.getAsync());
 Settings.onChange((s) => settings.set(s));
@@ -19,19 +17,6 @@ await notifications.init();
 let notificationsMode = lazyRune(() => invoke(SeelenCommand.GetNotificationsMode));
 subscribe(SeelenEvent.NotificationsModeChanged, notificationsMode.setByPayload);
 await notificationsMode.init();
-
-$effect.root(() => {
-  $effect(() => {
-    console.log("notifications: ", notifications.value);
-    console.log("notificationsMode: ", notificationsMode.value);
-  });
-});
-
-widget.window.onFocusChanged((e) => {
-  if (!e.payload) {
-    widget.hide();
-  }
-});
 
 class State {
   get notifications(): AppNotification[] {